feat: ゲストとログイン時のマップ画面を統合する
未ログインでも /map を表示し、ボトムパネルの3ボタンを常時表示する。
認証が必要な操作時のみ LoginModal を表示するようルーティングと UI を整理した。
