+++ client/src/routes/map/index.tsx
@@ -1,13 +1,10 @@
-import { createFileRoute, redirect } from "@tanstack/react-router";
+import { createFileRoute } from "@tanstack/react-router";
 import MapComponent from "@/components/map/map";
-import { getAccessToken } from "@/lib/apiClient";
 
 export const Route = createFileRoute("/map/")({
-  beforeLoad: () => {
-    if (!getAccessToken()) {
-      throw redirect({ to: "/login" });
-    }
-  },
+  validateSearch: (search: Record<string, unknown>) => ({
+    login: search.login === "1" || search.login === true,
+  }),
   component: MapComponent,
 });
